+++ public/header.php
@@ -1,0 +1,38 @@
+
+<!DOCTYPE HTML>
+
+<html>
+<head>
+    <script src="//code.jquery.com/jquery-1.11.1.min.js"></script>
+    <script src="//netdna.bootstrapcdn.com/bootstrap/3.1.0/js/bootstrap.min.js"></script>
+    <script type="text/javascript" src="./index.js"></script>
+    <title>SandwicherIIE</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
+    <link rel="stylesheet" href="assets/css/main.css" />
+    <noscript><link rel="stylesheet" href="assets/css/noscript.css" /></noscript>
+</head>
+<body  class="is-preload">
+<div id="wrapper">
+    <header id="header">
+        <div class="logo">
+            <img src="images/logo.png" width="80%"></img>
+        </div>
+        <div class="content">
+            <div class="inner">
+                <h1>La SandwicherIIE</h1>
+                <p>L'association Sandwich de l'ENSIIE, commandez votre menu avec nous</p>
+            </div>
+        </div>
+        <nav>
+            <ul>
+                <li><a href="">Connexion</a></li>
+                <li><a href="">Commander</a></li>
+				<li><a href="">Contact</a></li>
+				<li><a href="">Elements</a></li>
+            </ul>
+        </nav>
+    </header>
+
+    <!-- Main -->
+    <div id="main">
